+++ CHANGELOG.md
@@ -23,8 +23,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1034](https://github.com/spegel-org/spegel/pull/1034) Refactor to use shared reference in image and distribution path.
 - [#1037](https://github.com/spegel-org/spegel/pull/1037) Change to returning balancer from router lookups.
 - [#1039](https://github.com/spegel-org/spegel/pull/1039) Add caching for balancers to reduce duplicate queries.
+- [#1049](https://github.com/spegel-org/spegel/pull/1049) Refactor filters to support more than only regex.
+- [#1047](https://github.com/spegel-org/spegel/pull/1047) Refactor mirrored registries to be applied with registry filters.
 - [#1046](https://github.com/spegel-org/spegel/pull/1046) Add GrafanaDashboard option to Helm chart.
-
+  
 ### Deprecated
 
 ### Removed
@@ -42,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1032](https://github.com/spegel-org/spegel/pull/1032) Fix measuring pulls in debug web view with non default registry address configured.
 - [#1035](https://github.com/spegel-org/spegel/pull/1035) Implement single registry filter using reference.
 - [#1040](https://github.com/spegel-org/spegel/pull/1040) Fix distribution regex to support consecutive dashes and underscores.
+- [#1048](https://github.com/spegel-org/spegel/pull/1048) Implement filtering of content in state tracking.
 
 ### Security